@@ -7,7 +7,7 @@ import { ChangeDetectionStrategy, Component, Input, NgModule, ViewEncapsulation 
 @Component({
     selector: 'p-progressSpinner',
     template: `
-        <div class="p-progress-spinner" [ngStyle]="style" [ngClass]="styleClass" role="progressbar" [attr.aria-busy]="true" [attr.data-pc-name]="'progressspinner'" [attr.data-pc-section]="'root'">
+        <div class="p-progress-spinner" [ngStyle]="style" [ngClass]="styleClass" role="progressbar" [attr.aria-label]="ariaLabel" [attr.aria-busy]="true" [attr.data-pc-name]="'progressspinner'" [attr.data-pc-section]="'root'">
             <svg class="p-progress-spinner-svg" viewBox="25 25 50 50" [style.animation-duration]="animationDuration" [attr.data-pc-section]="'root'">
                 <circle class="p-progress-spinner-circle" cx="50" cy="50" r="20" [attr.fill]="fill" [attr.stroke-width]="strokeWidth" stroke-miterlimit="10" />
             </svg>
@@ -46,6 +46,11 @@ export class ProgressSpinner {
      * @group Props
      */
     @Input() animationDuration: string = '2s';
+    /**
+     * Used to define a aria label attribute the current element.
+     * @group Props
+     */
+    @Input() ariaLabel: string | undefined;
 }
 
 @NgModule({
