@@ -150,7 +150,15 @@ export class TabMenu implements AfterContentInit, AfterViewInit, AfterViewChecke
      * Defines the default active menuitem
      * @group Props
      */
-    @Input() activeItem: MenuItem | undefined;
+    @Input() set activeItem(value: MenuItem | undefined) {
+        this._activeItem = value;
+        this.activeItemChange.emit(value);
+        this.tabChanged = true;
+    }
+
+    get activeItem(): MenuItem | undefined {
+        return this._activeItem;
+    }
     /**
      * When enabled displays buttons at each side of the tab headers to scroll the tab list.
      * @group Props
@@ -221,6 +229,8 @@ export class TabMenu implements AfterContentInit, AfterViewInit, AfterViewChecke
 
     _model: MenuItem[] | undefined;
 
+    _activeItem: MenuItem | undefined;
+
     focusedItemInfo = signal<any>(null);
 
     get focusableItems() {
@@ -267,7 +277,7 @@ export class TabMenu implements AfterContentInit, AfterViewInit, AfterViewChecke
     }
 
     ngAfterViewChecked() {
-        if (this.tabChanged) {
+        if (isPlatformBrowser(this.platformId) && this.tabChanged) {
             this.updateInkBar();
             this.tabChanged = false;
         }
