@@ -29,7 +29,7 @@ import { Nullable } from 'primeng/ts-helpers';
                 <div class="p-timeline-event-opposite" [attr.data-pc-section]="'opposite'">
                     <ng-container *ngTemplateOutlet="oppositeTemplate; context: { $implicit: event }"></ng-container>
                 </div>
-                <div class="p-timeline-event-separator" [attr.data-pc-section]="'seperator'">
+                <div class="p-timeline-event-separator" [attr.data-pc-section]="'separator'">
                     <ng-container *ngIf="markerTemplate; else marker">
                         <ng-container *ngTemplateOutlet="markerTemplate; context: { $implicit: event }"></ng-container>
                     </ng-container>
