@@ -2,6 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { ButtonModule } from 'primeng/button';
+import { ButtonGroupModule } from 'primeng/buttongroup';
 import { RippleModule } from 'primeng/ripple';
 import { AppDocModule } from '../../layout/doc/app.doc.module';
 import { AppCodeModule } from '../../layout/doc/app.code.component';
@@ -25,9 +26,11 @@ import { SizesDoc } from './sizesdoc';
 import { StyleDoc } from './styledoc';
 import { TemplateDoc } from './templatedoc';
 import { TextDoc } from './textdoc';
+import { ButtonGroupDoc } from './buttongroupdoc';
+
 
 @NgModule({
-    imports: [CommonModule, AppCodeModule, AppDocModule, ButtonModule, RippleModule, RouterModule],
+    imports: [CommonModule, AppCodeModule, AppDocModule, ButtonModule, ButtonGroupModule, RippleModule, RouterModule],
     exports: [AppDocModule],
     declarations: [
         ImportDoc,
@@ -44,12 +47,13 @@ import { TextDoc } from './textdoc';
         OutlinedDoc,
         IconOnlyDoc,
         BadgeDoc,
+        ButtonGroupDoc,
         ButtonsetDoc,
         SizesDoc,
         TemplateDoc,
         StyleDoc,
         AccessibilityDoc,
-        DirectiveDoc
+        DirectiveDoc,
     ]
 })
 export class ButtonDocModule {}
