@@ -8,16 +8,16 @@ import { Code } from '../../domain/code';
             <p>An infinite spin animation is displayed by default.</p>
         </app-docsectiontext>
         <div class="card flex justify-content-center">
-            <p-progressSpinner aria-label="loading"></p-progressSpinner>
+            <p-progressSpinner ariaLabel="loading"></p-progressSpinner>
         </div>
         <app-code [code]="code" selector="progress-spinner-basic-demo"></app-code>
     `
 })
 export class BasicDoc {
     code: Code = {
-        basic: `<p-progressSpinner></p-progressSpinner>`,
+        basic: `<p-progressSpinner ariaLabel="loading"></p-progressSpinner>`,
         html: `<div class="card flex justify-content-center">
-    <p-progressSpinner></p-progressSpinner>
+    <p-progressSpinner ariaLabel="loading"></p-progressSpinner>
 </div>`,
         typescript: `
 import { Component } from '@angular/core';
