@@ -2,7 +2,6 @@ import { Component } from '@angular/core';
 import { AccessibilityDoc } from '../../doc/button/accessibilitydoc';
 import { BadgeDoc } from '../../doc/button/badgedoc';
 import { BasicDoc } from '../../doc/button/basicdoc';
-import { ButtonsetDoc } from '../../doc/button/buttonsetdoc';
 import { DirectiveDoc } from '../../doc/button/directivedoc';
 import { DisabledDoc } from '../../doc/button/disableddoc';
 import { IconsDoc } from '../../doc/button/iconsdoc';
@@ -19,7 +18,7 @@ import { SizesDoc } from '../../doc/button/sizesdoc';
 import { StyleDoc } from '../../doc/button/styledoc';
 import { TemplateDoc } from '../../doc/button/templatedoc';
 import { TextDoc } from '../../doc/button/textdoc';
-
+import { ButtonGroupDoc } from '../../doc/button/buttongroupdoc';
 @Component({
     templateUrl: './buttondemo.html',
     styleUrls: ['./buttondemo.scss']
@@ -102,9 +101,9 @@ export class ButtonDemo {
             component: BadgeDoc
         },
         {
-            id: 'buttonset',
-            label: 'Button Set',
-            component: ButtonsetDoc
+            id: 'buttongroup',
+            label: 'Button Group',
+            component: ButtonGroupDoc
         },
         {
             id: 'sizes',
