@@ -13,6 +13,7 @@ import { AccessibilityDoc } from '../../doc/dropdown/accessibilitydoc';
 import { ReactiveFormsDoc } from '../../doc/dropdown/reactiveformsdoc';
 import { LazyVirtualScrollDoc } from '../../doc/dropdown/lazyvirtualscrolldoc';
 import { InvalidDoc } from '../../doc/dropdown/invaliddoc';
+import { CustomFilterDoc } from '../../doc/dropdown/customfilterdoc';
 
 @Component({
     templateUrl: './dropdowndemo.html',
@@ -53,8 +54,20 @@ export class DropdownDemo {
         {
             id: 'filter',
             label: 'Filter',
-            component: FilterDoc
+            children: [
+                {
+                    id: 'filterbasic',
+                    label: 'Basic',
+                    component: FilterDoc
+                },
+                {
+                    id: 'customfilter',
+                    label: 'Custom Filter',
+                    component: CustomFilterDoc
+                }
+            ]
         },
+
         {
             id: 'virtualscroll',
             label: 'Virtual Scroll',
